+++ shinylive/_version/__init__.py
@@ -1,5 +1,5 @@
 # The version of this Python package.
-SHINYLIVE_PACKAGE_VERSION = "0.1.3.9000"
+SHINYLIVE_PACKAGE_VERSION = "0.2.0"
 
 # This is the version of the Shinylive assets to use.
-SHINYLIVE_ASSETS_VERSION = "0.2.4"
+SHINYLIVE_ASSETS_VERSION = "0.2.5"